Dispatcher auto-response buttons now on each tweet. TwitterFeed component now scrollable. Dividers now a part of tweet objects, rather than lists. Dummy data no longer in use.
